+++ demo/entities/src/main/java/software/xdev/sse/demo/entities/IdentifiableEntity.java
@@ -10,7 +10,7 @@ import jakarta.persistence.MappedSuperclass;
 @MappedSuperclass
 public abstract class IdentifiableEntity
 {
-	static final int DEFAULT_STRING_LENGTH = 255;
+	public static final int DEFAULT_STRING_LENGTH = 255;
 	
 	public static final String COL_ID = "id";